requirements.txt
google-generativeai
playwright
chromadb
pydub
speechrecognition
transformers
tqdm
pyttsx3